Updated routes, removed v1 blueprint, registered new blueprint for task related features
