+++ src/server/routes/admin.py
